Conditionally filter adaptive/responsive typography tokens
