+++ scripts/post-process-css/formatters/index.ts
@@ -1,3 +1,3 @@
 export { ConsolidateScreenSizes } from "./consolidate-screen-sizes/consolidate-screen-sizes";
-export { FilterAdaptiveTypography } from "./filter-adaptive-typography/filter-adaptive-typography";
+export { FilterTypographyTokens } from "./filter-typography-tokens/filter-typography-tokens";
 export { LightDarkModeFormatter } from "./light-dark-mode/light-dark-mode-formatter";
