+++ scripts/post-process-css/formatters/light-dark-mode/light-dark-mode-formatter.spec.ts
@@ -100,6 +100,7 @@ describe("LightDarkModeFormatter", () => {
 
       const result = lightDarkModeFormatter.formatTokens(tokens);
       expect(result).to.deep.equal({
+        context: "product",
         screenSizes: [
           {
             global: [],
